submodules/common_flask_frontend/requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.3.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-dotenv==1.0.0
SQLAlchemy==2.0.15
typing-extensions==4.5.0
Werkzeug==2.3.4
zipp==3.15.0
